+++ karma.conf.js
@@ -1,14 +1,15 @@
 /** @param {import('karma').Config} config */
 module.exports = config => config.set({
+
   client: {
     captureConsole: false,
     mocha: {
-      timeout: 100000
+      timeout: 10000000
     },
   },
-  browserDisconnectTimeout: 100000,
-  browserDisconnectTolerance: 10,
-  browserNoActivityTimeout: 100000,
+  browserDisconnectTimeout: 10000000,
+  browserDisconnectTolerance: 1000,
+  browserNoActivityTimeout: 10000000,
   flags: [
     '--disable-gpu',
     '--no-sandbox'