+++ src/components/InstancesList/InstancesList.js
@@ -580,7 +580,11 @@ class InstancesList extends React.Component {
           />
         </CheckboxColumn>
       ),
-      'title': ({ title }) => (
+      'title': ({
+        title,
+        discoverySuppress,
+        staffSuppress,
+      }) => (
         <AppIcon
           size="small"
           app="inventory"
@@ -588,6 +592,15 @@ class InstancesList extends React.Component {
           iconAlignment="baseline"
         >
           {title}
+          {(discoverySuppress || staffSuppress) &&
+          <span className={css.warnIcon}>
+            <Icon
+              size="medium"
+              icon="exclamation-circle"
+              status="warn"
+            />
+          </span>
+          }
         </AppIcon>
       ),
       'relation': r => formatters.relationsFormatter(r, data.instanceRelationshipTypes),