+++ src/settings/TargetProfileDetail.js
@@ -29,6 +29,12 @@ class TargetProfileDetail extends React.Component {
 
   render() {
     const { initialValues } = this.props;
+    const t = initialValues.targetOptions;
+    const targetOptions = !t ? null : (
+      <ul>
+        {Object.keys(t).sort().map(key => <li key={key}>{key}: {t[key]}</li>)}
+      </ul>
+    );
 
     return (
       <div>
@@ -75,8 +81,16 @@ class TargetProfileDetail extends React.Component {
         <Row>
           <Col xs={12}>
             <KeyValue
-              label={<FormattedMessage id="ui-inventory.jobProfileId" />}
-              value={initialValues.jobProfileId}
+              label={<FormattedMessage id="ui-inventory.createJobProfileId" />}
+              value={initialValues.createJobProfileId}
+            />
+          </Col>
+        </Row>
+        <Row>
+          <Col xs={12}>
+            <KeyValue
+              label={<FormattedMessage id="ui-inventory.updateJobProfileId" />}
+              value={initialValues.updateJobProfileId}
             />
           </Col>
         </Row>
@@ -84,7 +98,7 @@ class TargetProfileDetail extends React.Component {
           <Col xs={12}>
             <KeyValue
               label={<FormattedMessage id="ui-inventory.targetOptions" />}
-              value={initialValues.targetOptions}
+              value={targetOptions}
             />
           </Col>
         </Row>