+++ CHANGELOG.md
@@ -49,7 +49,7 @@
 * Add option to remove leading zeroes from Inventory HRIDs. Refs UIIN-1398.
 * Add ability to switch between new item statuses. Refs UIIN-1305.
 * Add option to remove leading zeroes from Inventory HRIDs. Refs UIIN-1398
-* Add support for maintaining copy-cataloguing profiles in the settings area. Towards UIIN-1401. Entails adding new interface `copycat-profiles`.
+* Add support for maintaining copy-cataloguing profiles in the settings area. Fixes UIIN-1401. Entails adding new interface `copycat-profiles`.
 * Add support for importing MARC records from external Z39.50 sources. Fixes UUIN-1317 and UUIN-1318. Entails adding new interface `copycat-imports`.
 * When multiple Z39.50 services are configured, offer a choice at import time. Fixes UIIN-1426.
 * Modify language of Import Record modal when updating a record. Fixes UIIN-1424.
@@ -57,6 +57,7 @@
 * Modify language of Settings/Inventory/Single record import. Fixes UIIN-1429.
 * Disable the single-record-import modal's Import button until something has been entered. Fixes UIIN-1431.
 * Better settings UI for maintaining copycat targets. Fixes UIIN-1437.
+* In "Z39.50 target profiles" settings, support the `targetOptions` field. Fixes UIIN-1445.
 * Update to `stripes-cli v2.0.0`. Refs UIIN-1410.
 * Show new request option for item with status restricted. Refs UIIN-1306.
 * Connect the `Remove leading zeroes from HRID` field to the `HRID handling` form. Refs UIIN-1413.
@@ -69,6 +70,12 @@
 * Add `tags.collection.get` as subpermission to `Inventory: View instances, holdings, and items` permission. Fixes UIIN-1422.
 * Display shelving order on the item record. Refs UIIN-816.
 * Passed the same ID down to the instance details pane and the loading pane. Fixes UIIN-1384.
+* Add item counter to each holding record. Refs UIIN-803.
+* Fix instance format filter. Refs UIIN-1423.
+* Change label Duplicate MARC bib record to Derive new MARC bib record. Refs UIIN-1436.
+* Add a warning icon for instance/holdings/item marked as Suppressed from discovery. Refs UIIN-1380.
+* Fix nature of content filter. Fixes UIIN-1441.
+* Add a warning icon for instance marked as Staff suppressed. Refs UIIN-1381.
 
 ## [5.0.1](https://github.com/folio-org/ui-inventory/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v5.0.0...v5.0.1)